Fix SonarCloud issues: bugs, security vulnerabilities, regex ReDoS
- Bugs: make PageManager.writeLock readonly (S2445, 15x), fix integer overflow in CompiledQueryExecutor (S3949), replace floating-point equality (S1244, 6x), remove dead code/unreachable checks (S2583), fix collection overwrite in LouvainAlgorithm (S4143), remove redundant Clear (S4158), check ReadAsync bytes read (S2674), dispose CancellationTokenSources (S2930), remove identical branches (S3923).

- Security: add path-traversal input validation to TenantBackupRestoreService (S2083/S6549), add regex execution timeouts to ~120 regex sites to prevent ReDoS (S6444), suppress non-security Random usage with justified NOSONAR comments (S2245, src).

- Document verified SonarC# false positives (local-function side effects in graph traversal/QueryCompiler) with NOSONAR comments.

- Add scripts/sonarcloud-overview.ps1 to read org/project health via the SonarCloud API.

Verified: SharpCoreDB.CI.slnf and benchmark projects build with 0 errors; all 1397 SharpCoreDB.Tests pass.

+++ src/SharpCoreDB/Services/CompiledQueryExecutor.cs
@@ -132,7 +132,7 @@ public class CompiledQueryExecutor
         {
             window = new List<(Dictionary<string, object>, IndexedRowData)>(
                 Math.Min(Math.Max(indexedRows.Count - offset, 0), limit));
-            var end = Math.Min(offset + limit, indexedRows.Count);
+            var end = (int)Math.Min((long)offset + limit, indexedRows.Count);
 
             for (int i = offset; i < end; i++)
             {
@@ -243,7 +243,7 @@ public class CompiledQueryExecutor
         {
             // Only allocate if OFFSET/LIMIT applied
             results = new List<Dictionary<string, object>>(Math.Min(filtered.Count - offset, limit));
-            var end = Math.Min(offset + limit, filtered.Count);
+            var end = (int)Math.Min((long)offset + limit, filtered.Count);
 
             for (int i = offset; i < end; i++)
             {
